Added the feature to create an account and verify the user using JWT

+++ BACKEND/Controllers/productController.js
@@ -1,6 +1,10 @@
 import pool from "../config/db.js";
 
 const getAllProducts = async (req, res) => {
+    const id_user = Number(req.user.id)
+    if (!id_user) {
+        return res.status(401).json({ message: 'User id not found: unauthorized user ' })
+    }
     try {
         const [rows] = await pool.query(`
             SELECT
@@ -19,12 +23,16 @@ const getAllProducts = async (req, res) => {
                 LEFT JOIN categories c ON p.category_id = c.id
                 LEFT JOIN suppliers s ON p.supplier_id = s.id
                 LEFT JOIN product_details pd ON p.id = pd.product_id
+                WHERE p.id_user = ? 
                 ORDER BY p.id DESC
-            `)
-        res.json(rows)
+            `, [id_user])
+
+        return res.status(200).json(rows);
     } catch (err) {
         console.error('getAllProducts error:', err)
-        res.status(500).json({ message: 'Failde to fetch products.' })
+        return res.status(500).json({
+            message: 'Failed to fetch products. Please try again later.'
+        });
     }
 }
 
@@ -32,17 +40,22 @@ const getAllProducts = async (req, res) => {
 const createProduct = async (req, res) => {
     const { name, price, quantity, category_id, supplier_id, description, image_url } = req.body
 
-    if (!name || price === undefined || quantity === undefined) {
-        return res.status(400).json({ message: 'name, price, and quantity are required.' })
+    const id_user = Number(req.user.id)
+
+    if (!id_user) {
+        return res.status(401).json({ message: 'User id not found: unauthorized user ' })
     }
 
-    const connection = await pool.getConnection()
+    let connection;
     try {
+        connection = await pool.getConnection()
         await connection.beginTransaction();
+
         const [productResult] = await connection.query(`
-                INSERT INTO products (name, price, quantity, category_id, supplier_id) VALUES (?, ?, ?, ?, ?)`,
-            [name, price, quantity, category_id || 1, supplier_id || 1]
+                INSERT INTO products (name, price, quantity, category_id, supplier_id, id_user) VALUES (?, ?, ?, ?, ?, ?)`,
+            [name, price, quantity, category_id || 1, supplier_id || 1, id_user]
         )
+
         const newProductId = productResult.insertId
 
         await connection.query(`
@@ -52,34 +65,43 @@ const createProduct = async (req, res) => {
 
         await connection.commit()
 
-        res.status(201).json({ message: 'Product created', id: newProductId })
+        return res.status(201).json({ message: 'Product created successfully', id: newProductId })
     } catch (err) {
-        await connection.rollback()
+        if (connection) await connection.rollback()
         console.error('createProcut error:', err)
-        res.status(500).json({ message: 'Failde to create product.' })
+        return res.status(500).json({ message: 'Internal Server Error: Failed to create product.' })
     } finally {
-        connection.release()
+        if (connection) connection.release()
     }
 }
 
-const updateProduct = async (req, res) => {
-    const { id } = req.params
-    const { name, price, quantity, category_id, supplier_id, description, image_url } = req.body
-    // ! start: problime if user update somting else not the three importat one code breacks
-    // if (!name || price === undefined || quantity === undefined) {
-    //     return res.status(400).json({ message: 'name, price, and quantity are required.' })
-    // }
-    // ! end: problime if user update somting else not the three importat one code breacks
 
-    const connection = await pool.getConnection()
+
+const updateProduct = async (req, res) => {
+    const { name, price, quantity, category_id, supplier_id, description, image_url } = req.body
+
+    const { id: productId } = req.params
+    const id_user = Number(req.user.id)
+
+    if (!id_user) {
+        return res.status(401).json({ message: 'User id not found: unauthorized user ' })
+    }
+    
+    let connection;
     try {
+        connection = await pool.getConnection()
         await connection.beginTransaction()
 
-        await connection.query(`
+        const [result] = await connection.query(`
                 UPDATE products SET name = ?, price = ?, quantity = ?, category_id = ?, supplier_id = ?
                 WHERE id = ?
-            `, [name, price, quantity, category_id || 1, supplier_id || 1, id]
+                AND id_user = ?
+            `, [name, price, quantity, category_id || 1, supplier_id || 1, productId, id_user]
         )
+
+        if (result.warningStatus === 0 && result.affectedRows === 0) {
+            return res.status(404).json({ message: `product not found: No product with the id:${productId} is available for this user.` })
+        }
 
         await connection.query(`
                 INSERT INTO product_details (product_id, description, image_url)
@@ -87,38 +109,45 @@ const updateProduct = async (req, res) => {
                 ON DUPLICATE KEY UPDATE
                     description = VALUES(description),
                     image_url = VALUES(image_url)
-            `, [id, description || null, image_url || null]
+            `, [productId, description || null, image_url || null]
         )
 
         await connection.commit()
-        res.status(201).json({ message: 'Product updated.' })
+        res.status(201).json({ message: 'Product updated successfully.' })
     } catch (err) {
-        await connection.rollback()
+        if (connection) await connection.rollback()
         console.error('updateProduct error:', err);
-        res.status(500).json({ message: 'Failde to update product' })
+        res.status(500).json({ message: 'Internal Server Error: Failde to update product' })
     } finally {
-        connection.release()
+        if (connection) connection.release()
     }
 }
 
 const deleteProduct = async (req, res) => {
-    const {id} = req.params
+    const { id: productId } = req.params
+    const id_user = Number(req.user.id)
+
+    if (!id_user) {
+        return res.status(401).json({ message: 'User id not found: unauthorized user ' })
+    }
+
     try {
         const [result] = await pool.query(`
                 DELETE FROM products
                 WHERE id = ?
-            `, [id]
+                AND id_user = ?
+            `, [productId, id_user]
         )
 
-        if (result.effectedRows === 0) {
-            return res.status(404).json({message: 'product not found'})
+        if (result.warningStatus === 0 && result.affectedRows === 0) {
+            return res.status(404).json({ message: `product not found: No product with the id:${productId} is available for you` })
         }
 
-        res.status(204).send()
+        return res.status(200).json({ message: 'Product was successfully deleted.' })
     } catch (err) {
         console.error('deleteProduct error:', err);
-        res.status(500).json({ message: 'Failde to delete product' })
+        return res.status(500).json({ message: 'Internal Server Error: Failde to delete product' })
     }
 }
 
-export { getAllProducts, createProduct, updateProduct, deleteProduct}
+export { getAllProducts, createProduct, updateProduct, deleteProduct }
